% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pub.R
\name{theme_pub}
\alias{theme_pub}
\title{theme_pub}
\usage{
theme_pub(base_size = 14, base_family = "Helvetica")
}
\arguments{
\item{base_size}{Font size for graphs. Defaults to 14}

\item{base_family}{Font for graphs. Defaults to Helvetica}
}
\description{
theme_pub
}
\examples{
data(iris)
ggplot(iris, aes(x=Species, y=Sepal.Length)) + geom_point() + theme_pub()
}
